+++ src/modules/ApiServices.mod.js
@@ -1,6 +1,5 @@
 import { involvementApiBaseURL, involvementApiAppId } from './variables.js';
 import Alert from './Alert.mod.js';
-import { displayComments } from './display.js';
 
 export default class ApiServices {
   static pushComment = (id, comment, username) => {
@@ -20,14 +19,13 @@ export default class ApiServices {
       .then((res) => {
         if (res.status === 201) {
           Alert.success('Comment added successfully', document.querySelector('.alert-container'));
-          displayComments(id);
         } else {
           Alert.error('Something went wrong', document.querySelector('.alert-container'));
         }
       });
   }
 
-  static getComments = async (id) => { 
+  static getComments = async (id) => {
     const url = `${involvementApiBaseURL}apps/${involvementApiAppId}/comments?item_id=${id}`;
     const res = await fetch(url);
     const data = await res.json();
