+++ src/modules/RenderElement.mod.js
@@ -1,3 +1,5 @@
+import ApiServices from './ApiServices.mod.js';
+
 export default class RenderElement {
   constructor(element, parent, position = 'beforeend', clearContent = false) {
     this.element = element;
@@ -14,3 +16,18 @@ export default class RenderElement {
     this.parent.insertAdjacentHTML(this.position, this.element);
   }
 }
+
+export const displayComments = (id) => {
+  const comments = ApiServices.getComments(id);
+  comments.then((data) => {
+    const commentsContainer = document.getElementById('comments-container');
+    commentsContainer.innerHTML = '';
+    data.forEach((comment) => {
+      commentsContainer.innerHTML += `
+      <div class="comment-${id}">
+        <p>${comment.creation_date} ${comment.username}: ${comment.comment} </p>
+      </div>
+      `;
+    });
+  });
+};
